Add exercises.

+++ ProgrammingFundamentalsExercises/DataTypesAndVariablesExercises/19.TheaThePhotographer/TheaThePhotographer.cs
@@ -14,10 +14,10 @@ class TheaThePhotographer
         long uploadTime = (long)goodPictures * uploadingTimeInSecondsPerGoodPicture;
         long totalTimeInSeconds = filterTime + uploadTime;
 
-        long days = totalTimeInSeconds / 86400;
-        long hours = totalTimeInSeconds % 86400 / 3600;
-        long minutes = totalTimeInSeconds % 86400 % 3600 / 60;
-        long seconds = totalTimeInSeconds % 86400 % 3600 % 60;
+        int days = (int)(totalTimeInSeconds / 86400);
+        int hours = (int)(totalTimeInSeconds % 86400 / 3600);
+        int minutes = (int)(totalTimeInSeconds % 86400 % 3600 / 60);
+        int seconds = (int)(totalTimeInSeconds % 86400 % 3600 % 60);
 
         Console.WriteLine("{0}:{1}:{2}:{3}", days, hours.ToString().PadLeft(2, '0'), minutes.ToString().PadLeft(2, '0'), seconds.ToString().PadLeft(2, '0'));
     }
